Use MMark, Lucid
Use MMark instead of Cheapskate for markdown parsing; use Lucid
instead of Blaze for HTML building. The point is that we can embed
our SiteM reader monad in Lucid's HtmlT monad transformer, which
makes some of the wiring much cleaner. I also like Lucid's Html DSL
better; and MMark for its part extracts the YAML metadata block for
me, so that's a bunch of code I get to delete, which is always
good.